Move IMUInverseKinematicsTool to tools library to avoid circular dependency and update references and registration accordingly, also update inheritance to remove duplication.
